+++ bible/src/main/java/com/library/bible/book/repository/IBookRepository.java
@@ -51,6 +51,6 @@ public interface IBookRepository {
 	List<Map<String, Object>> searchBooks(String keyword);
 	
 	List<Map<String, Object>> getBestSellerBookListMap();
-	int[] getBestSellerBookIdArray();
+	long[] getBestSellerBookIdArray();
 	
 }
